requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
langgraph==0.2.50
langchain==0.3.10
langchain-google-genai==2.0.5
python-dateutil==2.8.2
icalendar==5.0.11
aiosqlite==0.19.0
httpx==0.25.2
python-dotenv==1.0.0
